requirements.txt
streamlit
pandas
numpy
Pillow
pytorch
ultralytics
opencv-python
matplotlib
